rewire to make loading smarter
and not re-download the json on filtering
